Move EV-specific tests to cabf_ev
