+++ v2/lints/cabf_ev/lint_ev_organization_name_missing.go
@@ -1,4 +1,4 @@
-package cabf_br
+package cabf_ev
 
 /*
  * ZLint Copyright 2020 Regents of the University of Michigan
@@ -20,18 +20,18 @@ import (
 	"github.com/zmap/zlint/v2/util"
 )
 
-type evCountryMissing struct{}
+type evOrgMissing struct{}
 
-func (l *evCountryMissing) Initialize() error {
+func (l *evOrgMissing) Initialize() error {
 	return nil
 }
 
-func (l *evCountryMissing) CheckApplies(c *x509.Certificate) bool {
+func (l *evOrgMissing) CheckApplies(c *x509.Certificate) bool {
 	return util.IsEV(c.PolicyIdentifiers) && util.IsSubscriberCert(c)
 }
 
-func (l *evCountryMissing) Execute(c *x509.Certificate) *lint.LintResult {
-	if util.TypeInName(&c.Subject, util.CountryNameOID) {
+func (l *evOrgMissing) Execute(c *x509.Certificate) *lint.LintResult {
+	if util.TypeInName(&c.Subject, util.OrganizationNameOID) {
 		return &lint.LintResult{Status: lint.Pass}
 	} else {
 		return &lint.LintResult{Status: lint.Error}
@@ -40,11 +40,11 @@ func (l *evCountryMissing) Execute(c *x509.Certificate) *lint.LintResult {
 
 func init() {
 	lint.RegisterLint(&lint.Lint{
-		Name:          "e_ev_country_name_missing",
-		Description:   "EV certificates must include countryName in subject",
-		Citation:      "BRs: 7.1.6.1",
-		Source:        lint.CABFBaselineRequirements,
+		Name:          "e_ev_organization_name_missing",
+		Description:   "EV certificates must include organizationName in subject",
+		Citation:      "EVGs: 9.2.1",
+		Source:        lint.CABFEVGuidelines,
 		EffectiveDate: util.ZeroDate,
-		Lint:          &evCountryMissing{},
+		Lint:          &evOrgMissing{},
 	})
 }
