+++ v2/lints/cabf_ev/lint_ev_organization_name_missing_test.go
@@ -1,4 +1,4 @@
-package cabf_br
+package cabf_ev
 
 /*
  * ZLint Copyright 2020 Regents of the University of Michigan
@@ -21,19 +21,19 @@ import (
 	"github.com/zmap/zlint/v2/test"
 )
 
-func TestEvHasSN(t *testing.T) {
+func TestEvHasOrg(t *testing.T) {
 	inputPath := "evAllGood.pem"
 	expected := lint.Pass
-	out := test.TestLint("e_ev_serial_number_missing", inputPath)
+	out := test.TestLint("e_ev_organization_name_missing", inputPath)
 	if out.Status != expected {
 		t.Errorf("%s: expected %s, got %s", inputPath, expected, out.Status)
 	}
 }
 
-func TestEvNoSN(t *testing.T) {
-	inputPath := "evNoSN.pem"
+func TestEvNoOrg(t *testing.T) {
+	inputPath := "evNoOrg.pem"
 	expected := lint.Error
-	out := test.TestLint("e_ev_serial_number_missing", inputPath)
+	out := test.TestLint("e_ev_organization_name_missing", inputPath)
 	if out.Status != expected {
 		t.Errorf("%s: expected %s, got %s", inputPath, expected, out.Status)
 	}
