+++ v2/lints/cabf_ev/lint_ev_valid_time_too_long.go
@@ -1,4 +1,4 @@
-package cabf_br
+package cabf_ev
 
 /*
  * ZLint Copyright 2020 Regents of the University of Michigan
